Consume audio buffer and send to JS
Turns out to be on the critical path to getting boot. If we don't
pull data from the audio DMA channel it appears to fill up and the
boot process hangs very early.

We would have just had a no-op read, but might as well actually
send the data to the JS side. We use a 1khz polling loop, which is
perhaps too frequent, but it works for now.

The 6100 guitar chord startup chime now plays, but there is some
glitching after, and then a continous hum.

Polling for input also turns out to be on the critical path,
since we need to consume the input buffer to get the "audio context
running" flag sent from the browser process to the worker. We don't
actually do anything with the input data yet.