Improvements to JS display implementation
Fix red and green channels being swapped due to different assumptions
in channel ordering between DingusPPC and the JS side.

Add content hashing of the frame buffer so that we don't need to
blit to the JS side if the content hasn't changed. Uses xxHash3
since it has the best performance.