Ignore JS image file reads or writes before they're opened
Starting with dingusdev/dingusppc@446b1b8d996116b2db0d5361bc1c5c221ce9c2c6
the Gossamer ATA HD is read from when no image is specified, and the JS
implementation of ImgFile wsa throwing an exception (because disk 0 -
the default value - was not opened on the JS side). This changes
the class to ignore reads or writes if the file is not open.