Bump Emscripten fixed memory size to 512 + 32 MB
The Beige G3 was running into an Emscripten OOM because it could not
grow the heap when restarting while booting the 10.2 install CD with
no NVRAM contents (in which case the boot device is set the and whole
machine is sent a restart command).

I'm guessing this is due to heap fragmentation -- even though we free
the 256 MB allocated the first time the machine boots, we end up with
a different allocation pattern the second time around, and we thus can't
fit it in the same 256 MB zone. Allowing Emscripten to grow to 2x the
size is a workaround.