Add fast-path for flushing all PAT entries from the TLB
This happens very often during boot, and we were previously iterating
over all entries in the TLB to find ones with the right flags.

This patch adds a fast-path that just increments a sequence/version
counter. When checking for valid TLB entries, we can compare the
sequence counter to the one in the TLB entry.