Add multi-file disk image support.
MetaImgFile is a replacement for ImgFile.
It allows a virtual hard disk to be comprised of multiple parts of multiple files.
Each part is a called a chunk.
A chunk can be part of a file or can be bytes in RAM.

MetaImgFile::Open does the work of interpreting a disk image file. Currently, it supports raw disk images and Disk Copy 4.2 disk images.

If a disk image is a single partition with no partition map or a floppy disk image and is formatted with a known file system, then an Apple Partition Map is created in RAM.
Known file systems are MFS, HFS, and HFS+. The file system determines the partition type.

MetaImgFile::Open can append partitions from additional disk image files but only if all disk image files would create an Apple Partition Map virtual hard disk.

You can add as many partitions as you like up to 2 TB. Beyond that would require doubling the disk block size which might not be interpreted correctly in some cases.
Classic macOS seems to be limited to 16 or 18 partitions.