Tweak Emscripten linker flags to retain unmified exports and a name section
Required when moving to Emscripten 4.0.15 or later (https://github.com/emscripten-core/emscripten/blob/main/ChangeLog.md#4015---091725),
we now need to explicitly use `-g2` get a name section.

Something also changed around the same time that made the wasm
import/exports get minified, but we need `emscripten_date_now` and/or
`clock_time_get` to be accessible so we can fake the time if requested.
Per emscripten-core/emscripten#20762 the `-lexports.js` is the best way
to do ensure that the exports are not minified.